==> 2-Cleaning-data.ipynb <==
import pandas as pd
text = pd.read_csv("data/constitutions.csv")
labels = pd.read_csv("data/world-bank.csv")
text['country'] = text['country'].replace(['united-states-of-america'],'united-states')

countries_labels = set(dict.fromkeys(labels['country']))
countries_text = set(dict.fromkeys(text['country']))
negative_intersection = countries_labels.symmetric_difference(countries_text)

print(negative_intersection)

not_labeled = pd.DataFrame((text.loc[text['country'] =='saint-kitts-and-nevis']))
for elem in list(negative_intersection):
    not_labeled = not_labeled.append(text.loc[text['country'] == elem], ignore_index=False)
    text = text[text.country != elem]
    

countries_labels = set(dict.fromkeys(labels['country']))
countries_text = set(dict.fromkeys(text['country']))
negative_intersection = countries_labels.symmetric_difference(countries_text)

negative_intersection

import numpy as np

labels.replace("..", np.nan, inplace=True)

labels.to_csv("data/labels.csv", index=False)
labels.info()


labels.drop(columns=['Poverty headcount ratio at national poverty lines (% of population)',
                     'Poverty headcount ratio at $1.90 a day (2011 PPP) (% of population)',
                     'Income share held by lowest 20%',
                     'Contraceptive prevalence, any method (% of married women ages 15-49)',
                     'Prevalence of underweight, weight for age (% of children under 5)',
                     'Terrestrial and marine protected areas (% of total territorial area)',
                     'Electric power consumption (kWh per capita)',
                     'Energy use (kg of oil equivalent per capita)',
                     'Net lending (+) / net borrowing (-) (% of GDP)',
                     'Domestic credit provided by financial sector (% of GDP)',
                     'High-technology exports (% of manufactured exports)',
                     'Revenue, excluding grants (% of GDP)',
                     'Statistical Capacity Score (Overall Average) (scale 0 - 100)',
                     'Total debt service (% of exports of goods, services and primary income)',
                     'Tax revenue (% of GDP)',
                     'External debt stocks, total (DOD, current US$) (millions)',
                     
                    ], inplace=True)

labels.info()

#pd.set_option("display.max_rows", None, "display.max_columns", None)
text.sort_values(by = ['country', 'year'], ascending = [True, True], ignore_index=True, inplace=True)
text.head(10)

labels.head(10)

labels['constitution'] = ''

for index, row in labels.iterrows():
    
    labels.loc[index,'constitution'] = text['content'].loc[(text.country == row['country']) & (text.year < row['Date'])].max()

labels.info()

Los NaN de la columna 'constitución' se deben a que los países no poseían constitución antes de la fecha prevista. Se cambió NaN por string vacío.

labels['constitution'].fillna("", inplace=True)

labels.info()

labels.to_csv("data/merged.csv", index=False)
